OSX Port: file consolidation per code-review suggestion
