--- utils/getHostID_linux.go
@@ -1,32 +1,0 @@
-// Copyright 2016 The Serviced Authors.
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package utils
-
-import (
-	"os/exec"
-	"strings"
-)
-
-var hostIDCmdString = "/usr/bin/hostid"
-
-// getHostID retrieves the system's unique id, on linux this maps
-// to /usr/bin/hostid.
-func getHostID() (hostid string, err error) {
-	cmd := exec.Command(hostIDCmdString)
-	stdout, err := cmd.Output()
-	if err != nil {
-		return hostid, err
-	}
-	return strings.TrimSpace(string(stdout)), err
-}
